requirements.txt
tweepy
praw
requests
pillow >= 4